# eddist-admin/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "eddist-admin",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/auth/native/session": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Exchange OAuth2 access token for session token",
        "operationId": "post_native_session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NativeSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Session token created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NativeSessionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid access token"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "NativeSessionRequest": {
        "type": "object",
        "required": [
          "access_token"
        ],
        "properties": {
          "access_token": {
            "type": "string"
          }
        }
      },
      "NativeSessionResponse": {
        "type": "object",
        "required": [
          "session_token",
          "expires_at",
          "user_info"
        ],
        "properties": {
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "session_token": {
            "type": "string"
          },
          "user_info": {
            "$ref": "#/components/schemas/NativeUserInfo"
          }
        }
      },
      "NativeUserInfo": {
        "type": "object",
        "required": [
          "sub",
          "email",
          "preferred_username",
          "email_verified"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "email_verified": {
            "type": "boolean"
          },
          "preferred_username": {
            "type": "string"
          },
          "sub": {
            "type": "string"
          }
        }
      },
...
    }
  }
}